Tiny Marker and basic removal
When you zoom out to a given level, every polygon gets a little marker that can be used to drag it around (in case a particular shape is hard to see at that zoom level). This should be adaptive, but it not currently.

Right-click to remove a shape from the map (but it will be re-rendered at the next generation).