[unstable] Updates to movement
Now the shape doesn't need to be active to be moved. However, it's not great. Dragging this way causes the shape to teleport back to the start for one frame, then teleport to where it's supposed to be. In addition, dragging with the marker (this may have been a bug from teh start) after focusing the object with the marker has an issue: when dragging for the first time, the shape doesn't move until you let go of the button, then it teleports. Subsequent drags are totally fine though.