Magic Wand Implementation
Magic wand is workable now. It queries multiple locations, and can find stuff using nominatim and overpass. It still doesn't find containing areas (like if you click the center of a lake, it doesn't find the lake), but it's workable for small features like buildings, and you can still find lakes by clicking on the edge of the lake instead of the center.