Minor improvement to mobile experience
Markers are now always on by default when teh user is operating from a mobile device. Now, when the user's entire screen is an outline, they can still interface with the map. It doesn't drag the polygon. I quite like this approach, because it makes it a lot easier to use on mobile, and a lot clearer how to drag the shape.

In addition, some existing references to magicwand panel were removed.

If the user is on mobile, the control panel starts as closed.