Outline Quality added
You can now change the quality of new shapes being added to the map. It doesn't affect existing shapes yet. There are 4 options that affect the maximum number of coordinates for a shape: perfect sets the limit to infinity, great sets it to 5000 points, good to 1000 points, and low to 200 points.

Also, added Greenland.